requirements.txt
fastapi
strawberry-graphql
uvicorn
sqlalchemy
pydantic
